create hierarchical transfer tests
Creates lower-level test to do a proof of
concept test of the transfer API.  Also creates
a higher-level file that runs the lower-level test
over a sweep of file sizes and server configurations.

(Excepting this file, as it has some long filepaths
and long string commands that shouldn't be split up.
TEST_CHECKPATCH_SKIP_FILES=t/ci/200-stage-tests.sh

+++ t/ci/990-stop-server.sh
@@ -5,16 +5,23 @@
 
 test_description="Stopping the UnifyFS server"
 
-while [[ $# -gt 0 ]]
+CLEANUP_JOB_ON_EXIT="1"
+
+#while [[ $# -gt 0 ]]
+for arg in "$@"
 do
-    case $1 in
+    case $arg in
         -h|--help)
             ci_dir=$(dirname "$(readlink -fm $BASH_SOURCE)")
             $ci_dir/001-setup.sh -h
             exit
             ;;
-        *)
-            echo "usage ./990-stop-server.sh -h|--help"
+	--keep-job)
+	    CLEANUP_JOB_ON_EXIT="0"
+	    echo "Shut down servers but not end job"
+	    ;;
+	*)
+            echo "usage ./990-stop-server.sh -h|--help|--keep-job"
             exit 1
             ;;
     esac
@@ -34,32 +41,43 @@ test_expect_success "verify unifyfsd has stopped" '
     test_must_fail process_is_running unifyfsd 10
 '
 
-# If UNIFYFS_MOUNTPOINT is an existing dir, verify that is it empty
-test_expect_success REAL_MP "Verify UNIFYFS_MOUNTPOINT ($UNIFYFS_MP) is empty" '
-    test_dir_is_empty $UNIFYFS_MP
-'
 
-# Cleanup posix mountpoint
-test_expect_success POSIX "Cleanup UNIFYFS_CI_POSIX_MP: $UNIFYFS_CI_POSIX_MP" '
-    rm -rf $UNIFYFS_CI_POSIX_MP/*posix*
-'
+if [ "$CLEANUP_JOB_ON_EXIT" -eq "1" ]; then
 
-# cleanup_hosts
-test_expect_success PDSH,CLEAN "Cleanup hosts" '
-    cleanup_hosts
-'
-# Remove trap
-# If any tests failed, the suite will exit with 1 which will trigger the trap.
-# Since the hosts were already cleaned at this point, can remove trap to prevent
-# cleanup_hosts from being called again.
-trap - EXIT
 
-# end here if running tests individually
-if [[ -z $full_run ]]; then
-    ( test_done; )
-    test_exit_code=$?
+    # If UNIFYFS_MOUNTPOINT is an existing dir, verify that is it empty
+    test_expect_success REAL_MP \
+	  "Verify UNIFYFS_MOUNTPOINT ($UNIFYFS_MP) is empty" '
+        test_dir_is_empty $UNIFYFS_MP
+	'
 
-    cd "$(dirname "$SHARNESS_TRASH_DIRECTORY")"
+    # Cleanup posix mountpoint
+    test_expect_success POSIX \
+	  "Cleanup UNIFYFS_CI_POSIX_MP: $UNIFYFS_CI_POSIX_MP" '
+        rm -rf $UNIFYFS_CI_POSIX_MP/*posix*
+	'
 
-    return $test_exit_code
-fi
+    # cleanup_hosts
+    test_expect_success PDSH,CLEAN "Cleanup hosts" '
+        cleanup_hosts
+	'
+
+    # Remove trap
+    # If any tests failed, the suite will exit with 1 which
+    # will trigger the trap.  Since the hosts were already
+    # cleaned at this point, can remove trap to prevent
+    # cleanup_hosts from being called again.
+    trap - EXIT
+
+    # end here if running tests individually
+    if [[ -z $UNIFY_CI_TESTS_FULL_RUN ]]; then
+    	( test_done; )
+    	test_exit_code=$?
+
+    	cd "$(dirname "$SHARNESS_TRASH_DIRECTORY")"
+
+    	return $test_exit_code
+    fi
+
+fi # if [ $CLEANUP_ON_JOB_EXIT -eq "1" ]]; then
+
